requirements.txt
Django
djangorestframework
Pillow
psycopg2-binary
python-decouple
django-filter
drf_yasg
django-widget-tweaks